+++ src/main/java/codesquad/domain/deletehistory/ContentType.java
@@ -1,4 +1,4 @@
-package codesquad.domain;
+package codesquad.domain.deletehistory;
 
 public enum ContentType {
     ISSUE, ANSWER, LABEL
