+++ src/main/java/codesquad/domain/user/UserRepository.java
@@ -1,4 +1,4 @@
-package codesquad.domain;
+package codesquad.domain.user;
 
 import org.springframework.data.jpa.repository.JpaRepository;
