+++ src/main/java/codesquad/dto/UserDto.java
@@ -1,6 +1,6 @@
 package codesquad.dto;
 
-import codesquad.domain.User;
+import codesquad.domain.user.User;
 
 import javax.validation.constraints.Size;
