+++ src/main/java/codesquad/security/BasicAuthInterceptor.java
@@ -1,7 +1,7 @@
 package codesquad.security;
 
 import codesquad.UnAuthenticationException;
-import codesquad.domain.User;
+import codesquad.domain.user.User;
 import codesquad.service.UserService;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
