+++ src/main/java/codesquad/security/LoginUserHandlerMethodArgumentResolver.java
@@ -1,7 +1,7 @@
 package codesquad.security;
 
 import codesquad.UnAuthorizedException;
-import codesquad.domain.User;
+import codesquad.domain.user.User;
 import org.springframework.core.MethodParameter;
 import org.springframework.web.bind.support.WebDataBinderFactory;
 import org.springframework.web.context.request.NativeWebRequest;
