+++ src/main/java/codesquad/service/UserService.java
@@ -2,8 +2,8 @@ package codesquad.service;
 
 import codesquad.UnAuthenticationException;
 import codesquad.UnAuthorizedException;
-import codesquad.domain.User;
-import codesquad.domain.UserRepository;
+import codesquad.domain.user.User;
+import codesquad.domain.user.UserRepository;
 import codesquad.dto.UserDto;
 import org.springframework.stereotype.Service;
