+++ src/test/java/codesquad/web/UserAcceptanceTest.java
@@ -1,6 +1,6 @@
 package codesquad.web;
 
-import codesquad.domain.UserRepository;
+import codesquad.domain.user.UserRepository;
 import org.junit.Test;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
